Refactoring of  the main training loop
